change remove_overlap to fractions

+++ src/options.py
@@ -36,9 +36,9 @@ def add_data_arguments(parser):
                         help='project out previous models')
     group.add_argument('--remove', action='store_true',
                         help='remove examples that are predicted correctly by previous models')
-    group.add_argument('--remove-overlap', type=int, default=0,
+    group.add_argument('--remove-overlap', type=float, default=0,
                         help='Number of high overlapping examples to remove')
-    group.add_argument('--remove-random', type=int, default=0,
+    group.add_argument('--remove-random', type=float, default=0,
                         help='Number of random examples to remove')
     group.add_argument('--remove-cheat', type=str2bool, default=False,
                         help='remove examples where cheating features are enabled')
